news-service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.9.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
cssselect==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
gnews==0.3.6
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.1
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.8.1
orjson==3.10.3
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.4.28
requests==2.26.0
requests-file==2.0.0
rich==13.7.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
